requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
pydicom>=2.3.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
fhirpy>=1.3.0
docker-compose>=1.29.2